Move MEDIA settings to production.
Using the `static` function on urls.py raises an error if this setting
is not configured.

I could not read anything in the django_s3_storage documentation about
this setting, but it appears it overrides this variable anyway.

So now, using different storage backends for development and production
environments is working as expected.